requirements.txt

aiogram==3.21.0
aiosqlite==0.20.0
instagrapi==2.2.1
python-dotenv==1.0.1

emoji==1.7.0
Pillow==9.1.0
gspread==3.6.0
Unidecode==1.2.0
requests>=2.32.4
beautifulsoup4==4.9.3
pyTelegramBotAPI==4.5.1
google-api-python-client==1.10.0